extracted/moved isOutsideScreen function.
